% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_apps.R
\name{color_fct}
\alias{color_fct}
\title{color_fct}
\usage{
color_fct(variable, include_color_previews = T)
}
\arguments{
\item{variable}{a factor or character variable}

\item{include_color_previews}{boolean whether to include a preview table (takes a couple of seconds to render)}
}
\description{
color_fct
}
