% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_apps.R
\name{color_gg}
\alias{color_gg}
\title{color_gg}
\usage{
color_gg(gg_plot, include_color_previews = T)
}
\arguments{
\item{gg_plot}{a  gg plot}

\item{include_color_previews}{boolean whether to include a preview table (takes a couple of seconds to render)}
}
\description{
color_gg
}
