% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal_tableau}
\alias{pal_tableau}
\title{Tableau palette (10 or 20)}
\usage{
pal_tableau(n)
}
\arguments{
\item{n}{The number of colours (excluding an NA colour).}
}
\value{
A character vector of hex codes.
}
\description{
A function to retreive a vector of hex codes for a non-numeric (or non-ordererd) variable.
}
\examples{
scales::show_col(pal_tableau(15))
}
