% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_cols.R
\name{show_cols}
\alias{show_cols}
\title{Function to plot hex colors as barplot to visualize the colors}
\usage{
show_cols(hex_cols)
}
\arguments{
\item{hex_cols}{a character vector of hex colors}
}
\description{
Function to plot hex colors as barplot to visualize the colors
}
