% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_barplot.R
\name{vis_barplot}
\alias{vis_barplot}
\title{Create a quick and beautiful barplot}
\usage{
vis_barplot(
  df,
  x,
  y,
  fill = NULL,
  labels = T,
  digits = 2,
  scaling = function(x) {     return(x) },
  position = position_stack(),
  fontsize = 22,
  legend_title = "",
  labelsize = 3.5
)
}
\arguments{
\item{df}{A dataframe}

\item{x}{The X aesthetic}

\item{y}{The Y aesthetic}

\item{fill}{An optional fill aesthetic}

\item{labels}{Boolean to determine whether to include labels}

\item{digits}{Number of digits after the comma}

\item{scaling}{scaling function provided by the scales package or user defined functions}

\item{position}{position function provided by the ggplot2 package}

\item{fontsize}{base_size of the plot (applied via theme_bw(base_size=fontsize))}

\item{legend_title}{Label for the legend}

\item{labelsize}{Size of the labels if included}
}
\value{
A ggplot object
}
\description{
Create a quick and beautiful barplot
}
\examples{
tab <- diamonds\%>\%group_by(cut)\%>\%summarise(n=n())
vis_barplot(tab, x=cut, y=n, fill=cut, scaling=comma)
}
