% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_barplot.R
\name{vis_ranked_barplot}
\alias{vis_ranked_barplot}
\title{Bar ggplot.}
\usage{
vis_ranked_barplot(data, x, y, descending = TRUE, ...)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{x}{Unquoted x aesthetic variable.}

\item{y}{Unquoted y aesthetic variable, used for sorting}

\item{descending}{boolean to indicate whether to sort values in descending order}
}
\value{
A ggplot object.
}
\description{
Create a ranked bar plot
}
